+++ data-raw/border_colours.R
@@ -1,0 +1,8 @@
+border_colours <- c(
+  "#D62828",
+  "#F77F00",
+  "#2D6A4F",
+  "#495057"
+)
+
+usethis::use_data(border_colours, overwrite = TRUE)